% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{knit_report}
\alias{knit_report}
\title{Knit report within a targets pipeline}
\usage{
knit_report(rmd, output_format)
}
\arguments{
\item{rmd}{The Rmd file to be rendered}

\item{output_format}{The rmarkdown output format to be used}
}
\description{
Knit the report and store the rendered file in the shared output folder.
Adapted from Aurélien's \href{https://git-r3lab.uni.lu/aurelien.ginolhac/rworkshop}{rworkshop repository code} ()
}
