% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{tar_report}
\alias{tar_report}
\title{Knit the report and store the rendered file in the shared output folder.
A wrapper around `tarchetypes::tar_render()`}
\usage{
tar_report(name, rmd)
}
\arguments{
\item{name}{Symbol, name of the target}

\item{rmd}{The Rmd file of the report to be rendered}
}
\description{
Knit the report and store the rendered file in the shared output folder.
A wrapper around `tarchetypes::tar_render()`
}
