% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{write_heatmap_png}
\alias{write_heatmap_png}
\title{Save a ComplexHeatmap as a standalone png file}
\usage{
write_heatmap_png(
  x,
  filename,
  width = 500,
  height = 500,
  units = "px",
  res = 150
)
}
\arguments{
\item{x}{a ComplexHeatmap object}

\item{filename}{The basename of the output file}

\item{width, height, units, res}{The settings for the output file}
}
\description{
Save a ComplexHeatmap as a standalone png file
}
